refactor(kernel): move the page-side CDP forwarder out of the worker module
`startPageCdpForwarder` is page-only code; keeping it next to
`WorkerCdpProxy` meant the worker bundle's boot-critical graph nominally
carried it. It now lives in `cdp-page-forwarder.ts` (wire types stay in
`cdp-worker-proxy.ts`).

The first-load delta gate is raised 5 -> 8 kB for this change with the
per-module measurement recorded in `first-load-budget.json`: the growth is
the bridge session-model fix itself (registry, locks, self-heal, reset
signalling), not a hoisted chunk, and nothing lazy-loadable remains.
