+++ packages/webapp/first-load-budget.json
@@ -3,7 +3,8 @@
   "//delta": "maxDeltaKb is the PRIMARY gate: a PR may not grow either eager graph by more than this much versus its own merge-base, which the checker builds and measures on the same machine in the same run. Relative, so it fails only PRs that actually regress and never drifts with the platform. 4 kB was calibrated against real traffic, not guessed: eight in-flight PRs measured 0.00-1.54 kB, while a deliberately staged regression (a static `unpdf` import in the boot-critical supplemental-commands index) measured +7.0 kB. 4 kB sits in that gap.",
   "//ceiling": "The ceilings are the SECONDARY gate and the human-owned line in the sand. A delta gate alone lets many small under-threshold PRs creep the graph upward forever; these bound the total. They are NOT ratchets to nudge when a build goes red — raising one is a deliberate decision about cold-boot payload and needs a reason in the PR body. Tighten them freely as the graphs shrink.",
   "//delta-raised": "Raised 4 -> 5 for the biscotti guest-seat subsystem (PR #2535, 2026-08-28). NOT a hoist: the instruction markdown, the approver agent, and the whole directed-approval flow are behind first-use imports and verifiably absent from the worker bundle; the +4.2 kB that remains is code on the paths it gates (turn-gate state, the adaptTool hook, approval-router scoping). Every lazy-loading lever was taken first, and removing the boot wiring entirely bought more than byte-shaving did. This is a ceiling-style deliberate decision, not a ratchet to nudge: the next PR that comes in over 4 kB should be treated as a regression and investigated, not waved through, and this should be put back to 4 once the subsystem has landed.",
-  "maxDeltaKb": 5,
+  "//delta-raised-2417": "Raised 5 -> 8 for the CDP bridge session-model fix (issue #2417). NOT a hoist: measured per module (minified), the +7.6 kB is +5.0 kB in cdp/browser-api.ts (per-tab session registry, detach/LRU, stale-session self-heal, per-tab + bridge locks, session-scoped navigate), +0.8 kB cdp/cdp-client.ts (4002 upstream-reset handling, state-change notifications), +0.5 kB navigation-watcher.ts (own-session gating), +0.8 kB kernel/cdp-bridge.ts + cdp-worker-proxy.ts (cdp-reset / cdp-ready), +0.7 kB playwright handlers (session re-arm). All of it is code on the boot-critical bridge paths it gates; the page-only forwarder was split out of the worker graph (cdp-page-forwarder.ts) and nothing lazy-loadable remains. Same rule as above: treat the next PR over 5 kB as a regression, and put this back to 5 once #2417's follow-up (session-explicit handlers) lands.",
+  "maxDeltaKb": 8,
   "pageEagerCeilingKb": 768,
   "workerEagerCeilingKb": 4288
 }
